Fix metronome minus key wrapping
git-svn-id: https://svn.code.sf.net/p/ourorgan/svn/trunk@2160 69509848-2a73-49d5-a893-c4d8c934025a

+++ src/grandorgue/GOrgueMetronome.cpp
@@ -166,7 +166,7 @@ void GOrgueMetronome::ButtonChanged(int id)
 
 void GOrgueMetronome::UpdateBPM(int val)
 {
-	if (m_BPM + val < 1)
+	if ((int)m_BPM + val < 1)
 		m_BPM = 1;
 	else
 		m_BPM += val;
@@ -179,7 +179,7 @@ void GOrgueMetronome::UpdateBPM(int val)
 
 void GOrgueMetronome::UpdateMeasure(int val)
 {
-	if (m_MeasureLength + val < 0)
+	if ((int)m_MeasureLength + val < 0)
 		m_MeasureLength = 0;
 	else
 		m_MeasureLength += val;
